Add default REACT_APP_MAPBOX_ACCESS_TOKEN value
This makes it easier to get a working version set up. Because if you're new, you might not know this is even a variable that needs to be set in the first place.And when you do know it needs to be set, you have to fetch the name in one of the files, which is annoying.

Signed-off-by: JohnyTheCarrot <tuurmartens4@gmail.com>